backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
numpy==1.26.4
openai==1.59.6
python-multipart==0.0.20
requests==2.31.0
